fix(test): mv tests.
